Add support for keras model update instead of  retrain
These currently is not way to update a keras model ( calling fit )
without causing a new model to be build from scratch.

When `fit` is called in the KerasClassifier a new model object
is build. This may not be desired on the below stated online learning
paradigm.

This commit explores a pseudo online active learning pipeline
where instead of retraining the learner the model updates its weights
with the queried points.

This commit implements OnlineActiveLearning class, which simply monkey patches
the fit method on  the KerasClassifier object with underlying keras `model.fit`
method.